feature. 노션 기능 제거
