+++ .env.template
@@ -4,13 +4,5 @@ NEXT_PUBLIC_DESCRIPTION=
 NEXT_PUBLIC_DOMAIN=
 NEXT_PUBLIC_GITHUB=
 NEXT_PUBLIC_GITHUB_REPO= // `${string}/${string}` pattern
-NEXT_PUBLIC_TWITTER=
-NEXT_PUBLIC_GA_ID=
-
 NEXT_PUBLIC_ADMIN_EMAIL=
-
-NOTION_USER=
-NOTION_TOKEN=
-NOTION_TOKEN_V2=
-NOTION_BLOG_DATABASE_ID=
-NOTION_GUESTBOOK_PAGE_ID=
+NEXT_PUBLIC_GA_ID=
