--- src/@types/common.ts
@@ -1,6 +1,0 @@
-export type TypeModalKey = "login" | "confirm" | null;
-
-export interface IModal {
-  key: TypeModalKey;
-  func?: () => void;
-}
