+++ src/config/meta.ts
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import config from "@/config/notion.config";
+import config from "@/config/config";
 
 export function generateMetaTitle(subTitle: string) {
   return `${process.env.NEXT_PUBLIC_TITLE} :: ${subTitle}`;
